+++ odoo/local-src/stock_groupbypartner/tests/__init__.py
@@ -1,0 +1,2 @@
+from . import test_groupbypartner
+from . import test_propagate_carrier